Configure Redux store
