+++ src/reducers/imageFeed.js
@@ -1,0 +1,9 @@
+const initialState = {
+  currentFeed: 'cabin',
+  cabinCards: [],
+  architectureCards: []
+}
+
+export default function imageFeed(state = initialState, action) {
+  return state;
+}
