+++ src/reducers/index.js
@@ -1,0 +1,8 @@
+import { combineReducers } from 'redux';
+import imageFeed from './imageFeed';
+
+const rootReducer = combineReducers({
+  imageFeed
+});
+
+export default rootReducer;
